[Silver V] Title: 귀찮음, Time: 140 ms, Memory: 8176 KB, Score: 30 point -BaekjoonHub

+++ 백준/Silver/16208. 귀찮음/귀찮음.cc
@@ -7,8 +7,8 @@ int main()
 {
     int n;
     cin >> n;
-    vector<int> arr;
-    int sum = 0;
+    vector<long long> arr;
+    long long sum = 0;
     int val;
     for(int i = 0; i< n;++i)
     {
